requirements.txt
gradio>=3.50.2
biopython==1.79
mini3di==0.1.0
